ui/gradio_app/requirements.txt
gradio>=4.0.0
numpy
pandas
pillow
scikit-learn
timm
torch
torchvision
lightgbm
